We have no option but to use -Bsymbolic
pytorch 1.1 is built with a prehistoric compiler (gcc 4.9.2) that has
a different definition for some private stdlibc++ symbols.  Since the
compiler is so ancient, we can't use it to build the rest of our
stuff.

In addition to all that pytorch is build using the old c++ abi
(pre-c++11) and the rest of conda is built using the new one
(including our packages).

I really, really hope -Bsymbolic will work, otherwise we will have to
build our own pytorch packages, at least until 1.2.